Bug 1565268 - Recovering VRDisplay status after VRService is shutdown.
Summary:
When JS is paused, VRService will automatically shutdown because the timer is not updated,
we used to create a new VRDisplay to JS. However, the page is still using the older VRDisplay,
so it can't get the newest VRDisplayInfo status. We should exit presentation and
apply the VRDisplayInfo status to the original VRDisplay instead of creating new one.

Besides, we also need to release the VR screen buffer after exiting the immersive mode to avoid eglMakeCurrent error.

Reviewers: kip!

Tags: #secure-revision

Bug #: 1565268

Differential Revision: https://phabricator.services.mozilla.com/D50558